@@ -34,27 +34,27 @@
         "ng2-pica": "^1.2.0"
     },
     "peerDependencies": {
-        "@angular/core": ">=2.0.0",
-        "core-js": ">=2.4.1",
-        "rxjs": ">=5.0.0-beta.12"
+        "@angular/core": ">=6.0.0",
+        "core-js": ">=2.5.4",
+        "rxjs": ">=6.0.0"
     },
     "devDependencies": {
-        "@angular/common": ">=5.0.0",
-        "@angular/compiler": ">=5.0.0",
-        "@angular/compiler-cli": ">=5.0.0",
-        "@angular/core": ">=5.0.0",
-        "@angular/platform-browser": ">=5.0.0",
-        "@angular/platform-browser-dynamic": ">=5.0.0",
-        "@angular/router": ">=5.0.0",
+        "@angular/common": ">=6.0.0",
+        "@angular/compiler": ">=6.0.0",
+        "@angular/compiler-cli": ">=6.0.0",
+        "@angular/core": ">=6.0.0",
+        "@angular/platform-browser": ">=6.0.0",
+        "@angular/platform-browser-dynamic": ">=6.0.0",
+        "@angular/router": ">=6.0.0",
         "@types/node": ">=7.0.5",
-        "core-js": "^2.4.1",
-        "reflect-metadata": "^0.1.3",
+        "core-js": "^2.5.4",
+        "reflect-metadata": "^0.1.12",
         "rollup": "^0.41.4",
-        "rxjs": "5.1.1",
-        "tslint": "^4.4.2",
-        "typescript": "~2.4.2",
-        "uglify-js": "^2.7.5",
-        "zone.js": "^0.7.7"
+        "rxjs": "6.0.0",
+        "tslint": "^5.9.1",
+        "typescript": "~2.7.2",
+        "uglify-js": "^3.4.4",
+        "zone.js": "^0.8.26"
     },
     "bugs": {
         "url": "https://github.com/bergben/ng2-img-max/issues"
