@@ -1,6 +1,5 @@
 import { Injectable, Inject, forwardRef } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
+import { Subject, Observable } from 'rxjs';
 import { Ng2PicaService } from 'ng2-pica';
 
 import { ImgExifService } from './img-exif.service';
